Refactoring Game.java
